Add ui for patch release notes
