+++ src/main/java/io/gardenlinux/glvd/GlvdService.java
@@ -10,7 +10,10 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class GlvdService {
@@ -142,5 +145,45 @@ public class GlvdService {
 
     public List<DebSrc> sourcePackagesByGardenLinuxVersion(String version) {
         return debSrcRepository.findByDistId(Integer.parseInt(distVersionToId(version)));
+    }
+
+    private String getVersionByPackageName(List<DebSrc> input, String packageName) {
+        for (DebSrc debSrc : input) {
+            if (debSrc.getDebSource().equalsIgnoreCase(packageName)) {
+                return debSrc.getDebVersion();
+            }
+        }
+        return "";
+    }
+
+    public ReleaseNote releaseNote(final String gardenlinuxVersion) {
+        if (gardenlinuxVersion.endsWith(".0")) {
+            return new ReleaseNote(gardenlinuxVersion, List.of());
+        }
+        var v = new GardenLinuxVersion(gardenlinuxVersion);
+        var cvesNewVersion = getCveForDistribution(v.printVersion(), new SortAndPageOptions("cveId", "ASC", null, null));
+        var cvesOldVersion = getCveForDistribution(v.previousPatchVersion(), new SortAndPageOptions("cveId", "ASC", null, null));
+        var cvesNewVersionCveIds = cvesNewVersion.stream().map(SourcePackageCve::getCveId).collect(Collectors.joining());
+        var diff = cvesOldVersion.stream().filter(sourcePackageCve -> !cvesNewVersionCveIds.contains(sourcePackageCve.getCveId())).toList();
+        var packagesNew = sourcePackagesByGardenLinuxVersion(v.printVersion());
+        var packagesOld = sourcePackagesByGardenLinuxVersion(v.previousPatchVersion());
+        HashMap<String, List<String>> sourcePackageNameToCveListMapping = new HashMap<>();
+        for (SourcePackageCve sourcePackageCve : diff) {
+            var cveList = sourcePackageNameToCveListMapping.getOrDefault(sourcePackageCve.getSourcePackageName(), new ArrayList<>());
+            cveList.add(sourcePackageCve.getCveId());
+            sourcePackageNameToCveListMapping.put(sourcePackageCve.getSourcePackageName(), cveList);
+        }
+
+        List<ReleaseNotesPackage> releaseNotesPackages = new ArrayList<>();
+        sourcePackageNameToCveListMapping.forEach((sourcePackage, cves) ->
+                releaseNotesPackages.add(
+                        new ReleaseNotesPackage(sourcePackage,
+                                getVersionByPackageName(packagesOld, sourcePackage),
+                                getVersionByPackageName(packagesNew, sourcePackage),
+                                cves)
+                )
+        );
+
+        return new ReleaseNote(gardenlinuxVersion, releaseNotesPackages);
     }
 }
